Moved animation controls to animation tab.
